Straightened up use of removeBias calls.
Summary: The removeBias function take two arguments: T and TT. T is the variable type of the array. TT is the scalar type it consists of.

Test Plan: Compiles and no functional change.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D386

git-svn-id: https://svn.opendtect.org/od/trunk@29168 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/General/variogramcomputers.cc
@@ -319,7 +319,7 @@ bool VertVariogramComputer::compVarFromRange( DataPointSet& dpset, int colid,
 	    depth_out += (double)(step/zstep);
 	}
 
-	removeBias<double,float>( &interpolatedvals, &interpolatedvals, false );
+	removeBias<double,double>( &interpolatedvals, &interpolatedvals, false );
 	variogramvals_->set( nrcontribwells, 0, 0 );
 	axes_->set( nrcontribwells, 0, 0);
 	variogramnms_->add(grpnames.get( igroup-1 ));
